Prevent display of duplicate error messages
alertList gets appended to, regardless of alert type. In order to avoid
overwhelming the display we can prevent duplicate alerts from being
created/rendered.

Is this behavior that we actually want though? Without duplicate error
messages, inifinite loop situations may be harder to diagnose.

Also, the current implementation still triggers rerenders I think. Is
there a way to short circuit the AlertContainer so that it does not
attempt to update the state/rerender when there are just duplicate
alerts?